+++ lib/src/models/twitch_bits_leaderboard.dart
@@ -15,19 +15,19 @@ class TwitchBitsLeaderboard {
   final int score;
 
   TwitchBitsLeaderboard({
-    this.userId,
-    this.userLogin,
-    this.userName,
-    this.rank,
-    this.score,
+    required this.userId,
+    required this.userLogin,
+    required this.userName,
+    required this.rank,
+    required this.score,
   });
 
   factory TwitchBitsLeaderboard.fromJson(Map<String, dynamic> json) =>
       TwitchBitsLeaderboard(
-        userId: json['user_id'],
-        userLogin: json['user_login'],
-        userName: json['user_name'],
-        rank: json['rank'],
-        score: json['score'],
+        userId: json['user_id'] as String,
+        userLogin: json['user_login'] as String,
+        userName: json['user_name'] as String,
+        rank: json['rank'] as int,
+        score: json['score'] as int,
       );
 }