+++ lib/src/models/twitch_extension_transaction.dart
@@ -9,11 +9,11 @@ class _Cost {
   /// Always the string “Bits”.
   final String type;
 
-  _Cost({this.amount, this.type});
+  _Cost({required this.amount, required this.type});
 
   factory _Cost.fromJson(Map<String, dynamic> json) => _Cost(
-        amount: json['amount'],
-        type: json['type'],
+        amount: json['amount'] as int,
+        type: json['type'] as String,
       );
 }
 
@@ -30,13 +30,18 @@ class _ProductData {
   /// Flag used to indicate if the product is in development.
   final bool inDevelopment;
 
-  _ProductData({this.sku, this.cost, this.displayName, this.inDevelopment});
+  _ProductData({
+    required this.sku,
+    required this.cost,
+    required this.displayName,
+    required this.inDevelopment,
+  });
 
   factory _ProductData.fromJson(Map<String, dynamic> json) => _ProductData(
-        sku: json['sku'],
-        cost: _Cost.fromJson(json['cost']),
-        displayName: json['displayName'],
-        inDevelopment: json['inDevelopment'],
+        sku: json['sku'] as String,
+        cost: _Cost.fromJson(json['cost'] as Map<String, dynamic>),
+        displayName: json['displayName'] as String,
+        inDevelopment: json['inDevelopment'] as bool,
       );
 }
 
@@ -74,26 +79,26 @@ class TwitchExtensionTransaction {
   final _ProductData productData;
 
   /// Set this field to twitch.ext + your extension ID.
-  final String domain;
+  final String? domain;
 
   /// Flag that denotes whether or not the data was sent over the extension
   /// pubsub to all instances of the extension.
-  final bool broadcast;
+  final bool? broadcast;
 
   /// Always empty since only unexpired products can be purchased.
-  final String expiration;
+  final String? expiration;
 
   TwitchExtensionTransaction({
-    this.id,
-    this.timestamp,
-    this.broadcasterId,
-    this.broadcasterLogin,
-    this.broadcasterName,
-    this.userId,
-    this.userLogin,
-    this.userName,
-    this.productType,
-    this.productData,
+    required this.id,
+    required this.timestamp,
+    required this.broadcasterId,
+    required this.broadcasterLogin,
+    required this.broadcasterName,
+    required this.userId,
+    required this.userLogin,
+    required this.userName,
+    required this.productType,
+    required this.productData,
     this.domain,
     this.broadcast,
     this.expiration,
@@ -101,19 +106,20 @@ class TwitchExtensionTransaction {
 
   factory TwitchExtensionTransaction.fromJson(Map<String, dynamic> json) =>
       TwitchExtensionTransaction(
-        id: json['id'],
-        timestamp: DateTime.parse(json['timestamp']),
-        broadcasterId: json['broadcaster_id'],
-        broadcasterLogin: json['broadcaster_login'],
-        broadcasterName: json['broadcaster_name'],
-        userId: json['user_id'],
-        userLogin: json['user_login'],
-        userName: json['user_name'],
+        id: json['id'] as String,
+        timestamp: DateTime.parse(json['timestamp'] as String),
+        broadcasterId: json['broadcaster_id'] as String,
+        broadcasterLogin: json['broadcaster_login'] as String,
+        broadcasterName: json['broadcaster_name'] as String,
+        userId: json['user_id'] as String,
+        userLogin: json['user_login'] as String,
+        userName: json['user_name'] as String,
         productType:
             (json['product_type'] as String).toTransactionProductType(),
-        productData: _ProductData.fromJson(json['product_data']),
-        domain: json['domain'],
-        broadcast: json['broadcast'],
-        expiration: json['expiration'],
+        productData:
+            _ProductData.fromJson(json['product_data'] as Map<String, dynamic>),
+        domain: json['domain'] as String?,
+        broadcast: json['broadcast'] as bool?,
+        expiration: json['expiration'] as String?,
       );
 }