+++ lib/src/models/twitch_game_analytic.dart
@@ -7,17 +7,18 @@ class TwitchGameAnalytic {
   final String url;
 
   TwitchGameAnalytic({
-    this.gameId,
-    this.type,
-    this.url,
-    this.dateRange,
+    required this.gameId,
+    required this.type,
+    required this.url,
+    required this.dateRange,
   });
 
   factory TwitchGameAnalytic.fromJson(Map<String, dynamic> json) =>
       TwitchGameAnalytic(
-        gameId: json['game_id'],
-        type: json['type'],
-        url: json['URL'],
-        dateRange: TwitchDateRange.fromJson(json['date_range']),
+        gameId: json['game_id'] as String,
+        type: json['type'] as String,
+        url: json['URL'] as String,
+        dateRange: TwitchDateRange.fromJson(
+            json['date_range'] as Map<String, dynamic>),
       );
 }