+++ lib/src/models/twitch_search_category.dart
@@ -8,12 +8,16 @@ class TwitchSearchCategory {
   /// Game/category ID.
   final String id;
 
-  TwitchSearchCategory({this.boxArtUrl, this.name, this.id});
+  TwitchSearchCategory({
+    required this.boxArtUrl,
+    required this.name,
+    required this.id,
+  });
 
   factory TwitchSearchCategory.fromJson(Map<String, dynamic> json) =>
       TwitchSearchCategory(
-        boxArtUrl: json['box_art_url'],
-        name: json['name'],
-        id: json['id'],
+        boxArtUrl: json['box_art_url'] as String,
+        name: json['name'] as String,
+        id: json['id'] as String,
       );
 }