+++ lib/src/models/twitch_start_commercial.dart
@@ -8,12 +8,16 @@ class TwitchStartCommercial {
   /// Seconds until the next commercial can be served on this channel.
   final int retryAfter;
 
-  TwitchStartCommercial({this.length, this.message, this.retryAfter});
+  TwitchStartCommercial({
+    required this.length,
+    required this.message,
+    required this.retryAfter,
+  });
 
   factory TwitchStartCommercial.fromJson(Map<String, dynamic> json) =>
       TwitchStartCommercial(
-        length: json['length'],
-        message: json['message'],
-        retryAfter: json['retry_after'],
+        length: json['length'] as int,
+        message: json['message'] as String,
+        retryAfter: json['retry_after'] as int,
       );
 }