+++ lib/src/models/twitch_stream_info.dart
@@ -47,36 +47,40 @@ class TwitchStreamInfo {
   /// Shows tag IDs that apply to the stream.
   final List<String> tagIds;
 
+  final bool isMature;
+
   TwitchStreamInfo({
-    this.id,
-    this.userId,
-    this.userLogin,
-    this.userName,
-    this.gameId,
-    this.gameName,
-    this.type,
-    this.title,
-    this.viewerCount,
-    this.startedAt,
-    this.language,
-    this.thumbnailUrl,
-    this.tagIds,
+    required this.id,
+    required this.userId,
+    required this.userLogin,
+    required this.userName,
+    required this.gameId,
+    required this.gameName,
+    required this.type,
+    required this.title,
+    required this.viewerCount,
+    required this.startedAt,
+    required this.language,
+    required this.thumbnailUrl,
+    required this.tagIds,
+    required this.isMature,
   });
 
   factory TwitchStreamInfo.fromJson(Map<String, dynamic> json) =>
       TwitchStreamInfo(
-        id: json['id'],
-        userId: json['user_id'],
-        userLogin: json['user_login'],
-        userName: json['user_name'],
-        gameId: json['game_id'],
-        gameName: json['game_name'],
+        id: json['id'] as String,
+        userId: json['user_id'] as String,
+        userLogin: json['user_login'] as String,
+        userName: json['user_name'] as String,
+        gameId: json['game_id'] as String,
+        gameName: json['game_name'] as String,
         type: json['type'].toString().toStreamType(),
-        title: json['title'],
-        viewerCount: json['viewer_count'],
-        startedAt: DateTime.parse(json['started_at']),
-        language: json['language'],
-        thumbnailUrl: json['thumbnail_url'],
-        tagIds: List<String>.from(json['tag_ids']),
+        title: json['title'] as String,
+        viewerCount: json['viewer_count'] as int,
+        startedAt: DateTime.parse(json['started_at'] as String),
+        language: json['language'] as String,
+        thumbnailUrl: json['thumbnail_url'] as String,
+        tagIds: List<String>.from(json['tag_ids'] as Iterable),
+        isMature: json['is_mature'] as bool,
       );
 }